+++ NEWS.md
@@ -13,3 +13,4 @@
 * Fixed error where ace files with no plotspec field would not read in
 * Allow logical specification of antigens and sera for functions like `removeAntigens()`
 * Check you do not try and remove all antigens or sera
+* Add option to show map stress in plot